% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisonpictools-package.R
\docType{package}
\name{bisonpictools-package}
\alias{bisonpictools}
\alias{bisonpictools-package}
\title{bisonpictools: Ronald Wood Bison Camera Trap Tools}
\description{
This package facilitates the manipulation, exploration, and analysis of Ronald Wood Bison camera trap data.
}
\author{
\strong{Maintainer}: Nicole Hill \email{nicole@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7623-2153}{ORCID})

Other contributors:
\itemize{
  \item Andrea Kortello \email{andrea@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-8047-3331}{ORCID}) [contributor]
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
  \item Troy Hegel \email{Troy.Hegel@gov.ab.ca} [contributor]
  \item Alberta Environment and Parks [copyright holder]
}

}
\keyword{internal}
