% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-analyse.R
\name{bpt_analyse}
\alias{bpt_analyse}
\title{Fit Integrated Population Model}
\usage{
bpt_analyse(
  event_data,
  location_data,
  census,
  census_cv,
  census_study_year,
  census_day_of_year,
  proportion_calf,
  proportion_calf_cv,
  proportion_calf_study_year,
  proportion_calf_day_of_year,
  nthin = 1L,
  analysis_mode = "report"
)
}
\arguments{
\item{event_data}{tibble of templated event data}

\item{location_data}{tibble of templated location data}

\item{census}{integer vector of census values}

\item{census_cv}{numeric vector of census coefficients of variation}

\item{census_study_year}{character vector of census study years}

\item{census_day_of_year}{integer vector of census days of year}

\item{proportion_calf}{numeric vector of proportion of calf values}

\item{proportion_calf_cv}{numeric vector of proportion of calf coefficients
of variation}

\item{proportion_calf_study_year}{character vector of proportion of calf
study years}

\item{proportion_calf_day_of_year}{integer vector of proportion of calf days
of year}

\item{nthin}{thinning rate}

\item{analysis_mode}{character string of analysis mode, either "debug",
"check", or "report". Run firstin "debug" mode to ensure that the model
samples. Next, run in "report" mode to sample the full number of iterations.}
}
\value{
an \code{smb.analysis} object
}
\description{
Fits the integrated population model to data
}
\examples{
\dontrun{
bpt_analyse(
  event_data = bpt_event_data,
  location_data = bpt_location_data,
  census = 200L,
  census_cv = 0.1,
  census_study_year = "2021-2022",
  census_day_of_year = 365L,
  proportion_calf = c(0.201, 0.168),
  proportion_calf_cv = c(0.5, 0.5),
  proportion_calf_study_year = c("2020-2021", "2021-2022"),
  proportion_calf_day_of_year = c(365L, 365L),
  nthin = 1L,
  analysis_mode = "quick"
  )
}
}
