% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-bison-data.R
\name{bpt_check_data}
\alias{bpt_check_data}
\title{Data Checks}
\usage{
bpt_check_data(data, template)
}
\arguments{
\item{data}{A list containing the location and event data}

\item{template}{A list of the templates}
}
\value{
List of data frames
}
\description{
Confirms the columns types match the template, the values follow the allowed
ranges of the template and tables can be joined appropriately.
}
\examples{
bpt_check_data(
  list(
    Locations = location_data(),  
    Events = event_data()
   )
 )
}
