% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-check-study-year.R
\name{bpt_check_study_year}
\alias{bpt_check_study_year}
\title{Check Study Years}
\usage{
bpt_check_study_year(event_data, census_data, proportion_calf_data)
}
\arguments{
\item{event_data}{a tibble of templated event data}

\item{census_data}{a tibble of templated census data}

\item{proportion_calf_data}{a tibble of templated calf proportion data}
}
\value{
Throws an error if the study years do not match
}
\description{
Checks that the dates provided in the census and calf proportion data are
within the study years of the event data.
}
\examples{
bpt_check_study_year(
  bpt_event_data,
  bpt_census_data,
  bpt_proportion_calf_data
)
}
