% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-coefficient-table.R
\name{bpt_coefficient_table}
\alias{bpt_coefficient_table}
\title{Extract Coefficient Table
Extracts the coefficient table from the analysis object.}
\usage{
bpt_coefficient_table(analysis)
}
\arguments{
\item{analysis}{Analysis object from \code{bpt_analyse()}.}
}
\value{
a tibble of coefficients with columns:
\itemize{
\item \code{term} the name of the term
\item \code{estimate} the median posterior estimate
\item \code{lower} the lower 95\% credible interval
\item \code{upper} the upper 95\% credible interval
\item \code{svalue} a measure of directionality with respect to zero
\itemize{
\item an s-value of 4.32 bits is equivalent to a p-value of 0.05
}
}
}
\description{
Extract Coefficient Table
Extracts the coefficient table from the analysis object.
}
\examples{
\dontrun{
bpt_coefficient_table(analysis)
}
}
