% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-event-data.R
\name{bpt_event_data}
\alias{bpt_event_data}
\title{Generate example event data}
\usage{
bpt_event_data()
}
\value{
tibble of example event data
}
\description{
Generate example event data
}
\examples{
bpt_event_data()
}
