% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-load-analysis.R
\name{bpt_load_analysis}
\alias{bpt_load_analysis}
\title{Load analysis}
\usage{
bpt_load_analysis(file)
}
\arguments{
\item{file}{Path to file where analysis object is saved (excluding file
extension).}
}
\value{
An smb_analysis object.
}
\description{
Loads analysis object from file path where it has been saved.
}
\examples{
\dontrun{
bpt_load_analysis("file_path/analysis")
}
}
