% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-location-data.R
\name{bpt_location_data}
\alias{bpt_location_data}
\title{Generate example location data}
\usage{
bpt_location_data()
}
\value{
tibble of example location data
}
\description{
Generate example location data
}
\examples{
bpt_location_data()
}
