% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-location-matrix.R
\name{bpt_location_matrix}
\alias{bpt_location_matrix}
\title{Prepare location distance matrix for analysis}
\usage{
bpt_location_matrix(location_data)
}
\arguments{
\item{location_data}{Tibble of templated location data.}
}
\value{
A matrix of distances between locations in kilometers.
}
\description{
Calculates the distances (in km) between every location in the location_data
table.
}
\examples{
bpt_location_matrix(location_data)
}
