% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-manipulate-data-analysis-event.R
\name{bpt_manipulate_data_analysis_event}
\alias{bpt_manipulate_data_analysis_event}
\title{Prepare Event Data for Analysis}
\usage{
bpt_manipulate_data_analysis_event(event_data)
}
\arguments{
\item{event_data}{a tibble of templated event data}
}
\value{
a tibble of manipulated event data
}
\description{
Prepare Event Data for Analysis
}
\examples{
bpt_manipulate_data_analysis_event(bpt_event_data)
}
