% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-manipulate-data-analysis-location.R
\name{bpt_manipulate_data_analysis_location}
\alias{bpt_manipulate_data_analysis_location}
\title{Prepare Location Data for Analysis}
\usage{
bpt_manipulate_data_analysis_location(location_data)
}
\arguments{
\item{location_data}{a tibble of templated location data}
}
\value{
a tibble of manipulated location data
}
\description{
Prepare Location Data for Analysis
}
\examples{
bpt_manipulate_data_analysis_location(bpt_location_data)
}
