% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-manipulate-ratios.R
\name{bpt_manipulate_ratios}
\alias{bpt_manipulate_ratios}
\title{Manipulate Ratios}
\usage{
bpt_manipulate_ratios(
  data,
  numerator,
  denominator,
  study_years = unique(data$study_year),
  locations = unique(data$location_id)
)
}
\arguments{
\item{data}{A tibble of manipulated event, and location data.}

\item{numerator}{A character vector of sex-age codes to go in the numerator
of the ratio.}

\item{denominator}{A character vector of sex-age codes to go in the
denominator of the ratio.}

\item{study_years}{A character vector of study years to include in the plot.}

\item{locations}{A character vector of location_ids to include in the plot.}
}
\value{
The original tibble, with additional columns \code{numerator},
\code{denominator}, and \code{ratio}.
}
\description{
Takes output from \code{bpt_manipulate_data_plot()} and calculates ratio of
interest for plotting. Ratio of interest is calculated as
numerator:(denominator + numerator) to avoid infinite values.
}
\examples{
bpt_manipulate_ratios(
  data = bpt_manipulate_data_plot(
    bpt_event_data,
    bpt_location_data
  ),
  numerator = "fa",
  denominator = "ma",
  study_years = "2020-2021",
  locations = "LOCID1"
)
}
