% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-predict-abundance-class.R
\name{bpt_predict_abundance_class}
\alias{bpt_predict_abundance_class}
\title{Predict Abundance by Class}
\usage{
bpt_predict_abundance_class(analysis)
}
\arguments{
\item{analysis}{Analysis object from \code{bpt_analyse()}.}
}
\value{
A tibble of predicted abundances for each class and year, including
the median posterior value (estimate) and lower and upper 95\% credible
intervals.
}
\description{
Predicts abundances of each class for each study year.
}
\examples{
\dontrun{
bpt_predict_abundance_class(analysis)
}
}
