% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-predict-fecundity.R
\name{bpt_predict_fecundity}
\alias{bpt_predict_fecundity}
\title{Predict Fecundity}
\usage{
bpt_predict_fecundity(analysis)
}
\arguments{
\item{analysis}{Analysis object from \code{bpt_analyse()}.}
}
\value{
A tibble of predicted fecundity rate and proportion of reproductive
cows, including the median posterior value (estimate) and lower and upper
95\% credible intervals.
}
\description{
Predicts the fecundity rate and proportion of reproductive cows
}
\examples{
\dontrun{
bpt_predict_fecundity(analysis)
}
}
