% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-predict-ratios.R
\name{bpt_predict_ratios}
\alias{bpt_predict_ratios}
\title{Predict Ratios}
\usage{
bpt_predict_ratios(analysis)
}
\arguments{
\item{analysis}{Analysis object from \code{bpt_analyse()}.}
}
\value{
A tibble of predicted ratios for each study year, including the
median posterior value (estimate) and lower and upper 95\% credible
intervals.
}
\description{
Predicts the following ratios for each study year: M0:F0, M1:F1, Calf:FA,
Yearling:FA, M2:FA, M3:FA, MA:FA
}
\examples{
\dontrun{
bpt_predict_ratios(analysis)
}
}
