% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-predict-survival.R
\name{bpt_predict_survival}
\alias{bpt_predict_survival}
\title{Predict Survival}
\usage{
bpt_predict_survival(analysis)
}
\arguments{
\item{analysis}{Analysis object from \code{bpt_analyse()}.}
}
\value{
A tibble of predicted survival rates for each study year, including
the median posterior value (estimate) and lower and upper 95\% credible
intervals.
}
\description{
Predicts the survival rates for each class for each study
year. Bull survival corresponds to the M2, M3, and MA classes
}
\examples{
\dontrun{
bpt_predict_survival(analysis)
}
}
