% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-ratio-names.R
\name{bpt_ratio_names}
\alias{bpt_ratio_names}
\title{Names of common ratios}
\usage{
bpt_ratio_names(numerator, denominator)
}
\arguments{
\item{numerator}{A character vector of sex-age codes to go in the numerator
of the ratio.}

\item{denominator}{A character vector of sex-age codes to go in the
denominator of the ratio.}
}
\value{
A character vector of the common ratio name or "Ratio".
}
\description{
Names of common ratios
}
\examples{
# Common ratio
bpt_ratio_names(c("f0", "u0", "m0"), "fa")
# Uncommon ratio
bpt_ratio_names("u0", "fa")
}
