% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt_save_analysis.R
\name{bpt_save_analysis}
\alias{bpt_save_analysis}
\title{Save analysis object}
\usage{
bpt_save_analysis(analysis, file)
}
\arguments{
\item{analysis}{analysis object from \code{bpt_analyse()}}

\item{file}{file path to save analysis object to (file extension added
automatically)}
}
\description{
Saves analysis object to the file location specified by \code{file}.
}
\examples{
\dontrun{
bpt_save_analysis(analysis, "output/analysis")
}
}
