% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-seasons.R
\name{bpt_seasons}
\alias{bpt_seasons}
\title{Generate Seasons Data}
\usage{
bpt_seasons()
}
\value{
A tibble.
}
\description{
Generates a table of the dates of different seasons relevant to Wood
Bison movement patterns: calving, summer/fall, and winter. Start and end
dates of each season are based in the study year 1972-1973.
}
\examples{
bpt_seasons()
}
