% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-study-years.R
\name{bpt_study_years}
\alias{bpt_study_years}
\title{Calculate study years}
\usage{
bpt_study_years(data)
}
\arguments{
\item{data}{A tibble of templated data.}
}
\value{
a character vector of unique study years
}
\description{
Calculate the unique set of study years beginning on Apr. 1 of each year of a
templated \code{event_data}, \code{census_data}, or \code{proportion_calf_data} tibble.
}
\examples{
bpt_study_years(event_data)
}
