% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-census-data.R
\docType{data}
\name{census_data}
\alias{census_data}
\title{Sample data for census counts}
\format{
A tibble with 2 rows and 5 columns:
\describe{
\item{census_year}{The year of the census survey}
\item{census_month}{The month of the census survey}
\item{census_day}{The day of the census survey}
\item{census}{Estimated census count}
\item{census_cv}{Coefficient of variation (standard deviation / estimate)
of census estimate}
}
}
\usage{
census_data
}
\description{
A data set of invented census data.
}
\keyword{datasets}
