% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-event-data.R
\docType{data}
\name{event_data}
\alias{event_data}
\title{Sample Data for Camera Trap Events}
\format{
A tibble with 12 rows and 20 columns:
\describe{
\item{location_id}{The ID code for the location of the camera trap}
\item{start_year}{Year of start time of event}
\item{start_month}{Month of start time of event}
\item{start_day}{Day of start time of event}
\item{start_hour}{Hour of start time of event}
\item{start_minute}{Minute of start time of event}
\item{fa}{Number of female adults}
\item{f1}{Number of female yearlings}
\item{f0}{Number of female calves}
\item{fu}{Number of females of unknown age}
\item{ma}{Number of male adults}
\item{m3}{Number of male 3-year-olds}
\item{m2}{Number of male 2-year-olds}
\item{m1}{Number of male 1-year-olds}
\item{m0}{Number of male calves}
\item{mu}{Number of males of unknown age}
\item{ua}{Number of adults of unknown sex}
\item{u1}{Number of yearlings of unknown sex}
\item{u0}{Number of calves of unknown sex}
\item{uu}{Number of individuals of unknown age and unknown sex}
}
}
\usage{
event_data
}
\description{
A data set of invented camera trap event data
}
\keyword{datasets}
