% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location-data.R
\name{location_data}
\alias{location_data}
\title{Generate example location data}
\usage{
location_data()
}
\value{
tibble of example location data
}
\description{
Generate example location data
}
\examples{
location_data()
}
