% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-data-analysis-location.R
\name{manipulate_data_analysis_location}
\alias{manipulate_data_analysis_location}
\title{Prepare Location Data for Analysis}
\usage{
manipulate_data_analysis_location(location_data)
}
\arguments{
\item{location_data}{a tibble of templated location data}
}
\value{
a tibble of manipulated location data
}
\description{
Prepare Location Data for Analysis
}
\examples{
manipulate_data_analysis_location(location_data())
}
