% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-ratios.R
\name{manipulate_ratios}
\alias{manipulate_ratios}
\title{Manipulate Ratios}
\usage{
manipulate_ratios(
  data,
  numerator,
  denominator,
  years = unique(data$year),
  locations = unique(data$location_id)
)
}
\arguments{
\item{data}{a tibble of manipulated event and location data, with factor columns \code{year}, \code{location_id}, and integer columns c(\code{fa}, \code{f1}, \code{f0}, \code{fu}, \code{ma}, \code{m3}, \code{m2}, \code{m1}, \code{m0}, \code{mu}, \code{ua}, \code{u1}, \code{u0}, and \code{uu})}

\item{numerator}{a character vector of sex-age codes to go in the numerator of the ratio}

\item{denominator}{a character vector of sex-age codes to go in the denominator of the ratio}

\item{years}{a numeric vector of years to include in the plot}

\item{locations}{a character vector of location_ids to include in the plot}
}
\value{
the original tibble, with additional columns \code{numerator}, \code{denominator}, and \code{ratio}
}
\description{
Takes output from \code{manipulate_data_plot()} and calculates ratio of interest for plotting. Ratio of interest is calculated as numerator:(denominator + numerator) to avoid infinite values.
}
\examples{
manipulate_ratios(
 data = manipulate_data_plot(event_data(), location_data()),
 numerator = "fa", 
 denominator = "ma", 
 years = "2021", 
 locations = "RLBH007"
)
}
