% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ratios.R
\name{plot_ratios}
\alias{plot_ratios}
\title{Plot Ratios of Ronald Wood Bison camera events}
\usage{
plot_ratios(
  event_data,
  location_data,
  numerator,
  denominator,
  years = unique(event_data$start_year),
  locations = unique(location_data$location_id)
)
}
\arguments{
\item{event_data}{a tibble of templated event data}

\item{location_data}{a tibble of templated location data}

\item{numerator}{a character vector of sex-age codes to go in the numerator of the ratio}

\item{denominator}{a character vector of sex-age codes to go in the denominator of the ratio}

\item{years}{a numeric vector of years to include in the plot}

\item{locations}{a character vector of location_ids to include in the plot}
}
\value{
a \code{ggplot} object
}
\description{
Plot Ratios of Ronald Wood Bison camera events
}
\details{
The ratio is presented as numerator:(denominator + numerator)) to avoid infinite values when the numerator is ≥ 1 and the denominator = 0.
}
