% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-proportion-calf-data.R
\docType{data}
\name{proportion_calf_data}
\alias{proportion_calf_data}
\title{Sample data for calf proportion data}
\format{
A tibble with 2 rows and 5 columns:
\describe{
\item{proportion_calf_year}{The year of the calf proportion survey}
\item{proportion_calf_month}{The month of the calf proportion survey}
\item{proportion_calf_day}{The day of the calf proportion survey}
\item{proportion_calf}{Estimated calf proportion}
\item{proportion_calf_cv}{Coefficient of variation (standard deviation /
estimate) of calf proportion estimate}
}
}
\usage{
proportion_calf_data
}
\description{
A data set of invented calf proportion data.
}
\keyword{datasets}
