% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio-names.R
\name{ratio_names}
\alias{ratio_names}
\title{Names of common ratios}
\usage{
ratio_names(numerator, denominator)
}
\arguments{
\item{numerator}{a character vector of sex-age codes to go in the numerator of the ratio}

\item{denominator}{a character vector of sex-age codes to go in the denominator of the ratio}
}
\value{
character vector of common ratio name or "Ratio"
}
\description{
Names of common ratios
}
\examples{
# Common ratio
ratio_names(c("f0", "u0", "m0"), "fa")
# Uncommon ratio
ratio_names("u0", "fa")
}
