% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{seasons}
\alias{seasons}
\title{Generate Seasons Data}
\usage{
seasons()
}
\value{
A tibble
}
\description{
Generates a table of the dates of different seasons relevant to Ronald Wood Bison movement patterns: calving, summer/fall, and winter. Start and end dates of each season are based in the study year 1972-1973.
}
\examples{
seasons()
}
