% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisonpictools-package.R
\docType{package}
\name{bisonpictools-package}
\alias{bisonpictools}
\alias{bisonpictools-package}
\title{bisonpictools: Wood Bison Camera Trap Tools}
\description{
This package facilitates the manipulation, exploration, and analysis of wood bison camera trap data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/bisonpictools}
  \item \url{https://poissonconsulting.github.io/bisonpictools/}
  \item Report bugs at \url{https://github.com/poissonconsulting/bisonpictools/issues}
}

}
\author{
\strong{Maintainer}: Nicole Hill \email{nicole@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7623-2153}{ORCID})

Authors:
\itemize{
  \item Andrea Kortello \email{andrea@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-8047-3331}{ORCID})
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})
}

Other contributors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID}) [contributor]
  \item Troy Hegel \email{Troy.Hegel@gov.ab.ca} [contributor]
  \item Province of Alberta [copyright holder]
}

}
\keyword{internal}
