% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-analyse.R
\name{bpt_analyse}
\alias{bpt_analyse}
\title{Fit Integrated Population Model}
\usage{
bpt_analyse(
  event_data,
  location_data,
  census_data,
  proportion_calf_data,
  nthin = 10L,
  analysis_mode = "report"
)
}
\arguments{
\item{event_data}{Tibble of templated event data.}

\item{location_data}{Tibble of templated location data.}

\item{census_data}{Tibble of census data.}

\item{proportion_calf_data}{Tibble of proportion calf data.}

\item{nthin}{Thinning rate, takes default of 10.}

\item{analysis_mode}{Character string of analysis mode, either "debug",
"check", or "report".}
}
\value{
An \code{smb.analysis} object.
}
\description{
Fits the integrated population model to data. Run first in "debug" mode to
ensure that the model samples. Next, run in "report" mode to sample the
full number of iterations.
}
\examples{
\dontrun{
bpt_analyse(
  event_data = bpt_event_data,
  location_data = bpt_location_data,
  census_data = bpt_census_data,
  proportion_calf_data = bpt_proportion_calf_data,
  nthin = 1L,
  analysis_mode = "quick"
)
}
}
