% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data.R
\name{bpt_check_data}
\alias{bpt_check_data}
\title{Data Checks}
\usage{
bpt_check_data(data, template)
}
\arguments{
\item{data}{A list containing the location and event data}

\item{template}{A list of the templates}
}
\value{
Data
}
\description{
Data Checks
}
\examples{
bpt_check_data(
  list(
    Locations = location_data(),  
    Events = event_data()
   )
 )
}
