% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-load-analysis.R
\name{bpt_load_analysis}
\alias{bpt_load_analysis}
\title{Load Analysis}
\usage{
bpt_load_analysis(file)
}
\arguments{
\item{file}{path to file (excluding extension)}
}
\value{
the analysis object
}
\description{
Loads analysis object from file path where it has been saved
}
\examples{
\dontrun{
bpt_load_analysis("file_path/analysis")
}
}
