% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-location-matrix.R
\name{bpt_location_matrix}
\alias{bpt_location_matrix}
\title{Prepare Location Distance Matrix for Analysis}
\usage{
bpt_location_matrix(location_data)
}
\arguments{
\item{location_data}{a tibble of templated location data}
}
\value{
a matrix of distances between locations in kilometers
}
\description{
Prepare Location Distance Matrix for Analysis
}
\examples{
bpt_location_matrix(bpt_location_data)
}
