% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-manipulate-data-analysis.R
\name{bpt_manipulate_data_analysis}
\alias{bpt_manipulate_data_analysis}
\title{Prepare Data for Analysis}
\usage{
bpt_manipulate_data_analysis(
  event_data,
  location_data,
  census_data,
  proportion_calf_data
)
}
\arguments{
\item{event_data}{a tibble of templated event data}

\item{location_data}{a tibble of templated location data}

\item{census_data}{a tibble of templated census data}

\item{proportion_calf_data}{a tibble of templated proportion of calf data}
}
\value{
a list of manipulated data ready for analysis
}
\description{
Checks that data is in the correct format and manipulates it for analysis.
}
\examples{
bpt_manipulate_data_analysis(
  event_data = bpt_event_data,
  location_data = bpt_location_data,
  census_data = bpt_census_data,
  proportion_calf_data = bpt_proportion_calf_data
)
}
