% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-manipulate-data-plot.R
\name{bpt_manipulate_data_plot}
\alias{bpt_manipulate_data_plot}
\title{Manipulate Data for Plot}
\usage{
bpt_manipulate_data_plot(event_data, location_data)
}
\arguments{
\item{event_data}{Tibble of templated event data.}

\item{location_data}{Tibble of templated location data.}
}
\value{
A tibble of data prepared for plotting.
}
\description{
Manipulate Data for Plot
}
\examples{
bpt_manipulate_data_plot(
  event_data,
  location_data
)
}
