% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-plot-ratios.R
\name{bpt_plot_ratios}
\alias{bpt_plot_ratios}
\title{Plot Ratios of Ronald Wood Bison Camera Events}
\usage{
bpt_plot_ratios(
  event_data,
  location_data,
  numerator,
  denominator,
  study_years = bpt_study_years(event_data),
  locations = unique(location_data$location_id)
)
}
\arguments{
\item{event_data}{a tibble of templated event data}

\item{location_data}{a tibble of templated location data}

\item{numerator}{a character vector of sex-age codes to go in the numerator of the ratio}

\item{denominator}{a character vector of sex-age codes to go in the denominator of the ratio}

\item{study_years}{a character vector of study years to include in the plot}

\item{locations}{a character vector of location_ids to include in the plot}
}
\value{
a \code{ggplot} object
}
\description{
Generates bubble plot of ratio of the sex-age classes supplied to \code{numerator} and \code{denominator}. The ratio is given as \verb{numerator:(denominator + numerator))} to avoid infinite values when \verb{numerator ≥ 1} and \code{denominator = 0}. Each point represents an event. The size of the point represents the total group size, and the colour of the point represents the value of the ratio.
}
\examples{
# Plot calf:(cow + calf) ratio
bpt_plot_ratios(
  event_data = bpt_event_data(), 
  location_data = bpt_location_data(), 
  numerator = c("f0", "m0", "u0"), 
  denominator = c("fa")
 )
# Plot ratio of female:(male + female) of yearlings in 2021 at site RBLH007 only
bpt_plot_ratios(
  event_data = bpt_event_data(), 
  location_data = bpt_location_data(), 
  numerator = "f1", 
  denominator = "m1", 
  study_years = "2021-2022", 
  locations = "RLBH007"
 )
}
