% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpt-study-years.R
\name{bpt_study_years}
\alias{bpt_study_years}
\title{Calculate Study Years}
\usage{
bpt_study_years(event_data)
}
\arguments{
\item{event_data}{a tibble of templated event data}
}
\value{
a character vector of unique study years
}
\description{
Calculate the unique set of study years beginning on Apr. 1 of each year of a templated \code{event_data} tibble.
}
\examples{
bpt_study_years(bpt_event_data())
}
