% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-location-data.R
\docType{data}
\name{location_data}
\alias{location_data}
\title{Sample Data for Camera Trap Locations}
\format{
A tibble with 4 rows and 3 columns:
\describe{
\item{location_id}{The ID code for the location of the camera trap}
\item{latitude}{Latitude of camera trap location}
\item{longitude}{Longitude of camera trap location}
}
}
\usage{
location_data
}
\description{
A data set of invented camera trap locations
}
\keyword{datasets}
