% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-data-plot.R
\name{manipulate_data_plot}
\alias{manipulate_data_plot}
\title{Manipulate Data for Plot}
\usage{
manipulate_data_plot(event_data, location_data)
}
\arguments{
\item{event_data}{a tibble of templated event data}

\item{location_data}{a tibble of templated location data}
}
\value{
A tibble of data prepared for plotting.
}
\description{
Manipulate Data for Plot
}
\examples{
manipulate_data_plot(event_data(), location_data())
}
