% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ratios.R
\name{plot_ratios}
\alias{plot_ratios}
\title{Plot Ratios of Ronald Wood Bison camera events}
\usage{
plot_ratios(
  event_data,
  location_data,
  numerator,
  denominator,
  years = unique(as.character(event_data$start_year)),
  locations = unique(location_data$location_id)
)
}
\arguments{
\item{event_data}{a tibble of templated event data}

\item{location_data}{a tibble of templated location data}

\item{numerator}{a character vector of sex-age codes to go in the numerator of the ratio}

\item{denominator}{a character vector of sex-age codes to go in the denominator of the ratio}

\item{years}{a numeric vector of years to include in the plot}

\item{locations}{a character vector of location_ids to include in the plot}
}
\value{
a \code{ggplot} object
}
\description{
Generates bubble plot of ratio of the sex-age classes supplied to \code{numerator} and \code{denominator}. The ratio is given as numerator\verb{:(}denominator\code{+}numerator`)) to avoid infinite values when the numerator is ≥ 1 and the denominator = 0. Each point represents an event. The size of the point represents the total herd size, and the colour of the point represents the value of the ratio.
}
\examples{
# Plot calf:(cow + calf) ratio
plot_ratios(
  event_data = event_data(), 
  location_data = location_data(), 
  numerator = c("f0", "m0", "u0"), 
  denominator = c("fa")
 )
# Plot ratio of female:(male + female) of yearlings in 2021 at site RBLH007 only
plot_ratios(
  event_data = event_data(), 
  location_data = location_data(), 
  numerator = "f1", 
  denominator = "m1", 
  years = "2021", 
  locations = "RLBH007"
 )
}
