% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template}
\alias{template}
\title{Template for Entering Bison Data}
\format{
\subsection{\code{template}}{

A list with two data frames the Locations table contains:
\describe{
\item{location_id}{Location name}
\item{latitude}{Latitude of the location}
\item{longitude}{Longitude of the location}
...
}
}
}
\usage{
template
}
\description{
Template for Entering Bison Data
}
\keyword{datasets}
